Conflicts:
	src/main/java/net/jamcraft/chowtime/ChowTime.java
	src/main/java/net/jamcraft/chowtime/core/CTRegistry.java

	src/main/java/net/jamcraft/chowtime/core/gen/candyLand/WorldProviderCandyLand.java

+++ src/main/java/net/jamcraft/chowtime/core/events/EntityEventHandler.java
@@ -7,8 +7,15 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 import net.jamcraft.chowtime.ChowTime;
+import net.jamcraft.chowtime.core.crops.CropBarley;
+import net.minecraft.block.BlockDirt;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.nbt.CompressedStreamTools;
+import net.minecraft.util.ChatComponentText;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
+import net.minecraftforge.event.entity.living.LivingEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent;
+import net.minecraftforge.event.entity.player.PlayerUseItemEvent;
 import net.minecraftforge.event.world.WorldEvent;
 import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
@@ -18,8 +25,13 @@ public class EntityEventHandler
     @SubscribeEvent
     public void onEntityJoinWorld(EntityJoinWorldEvent event)
     {
+        if (event.entity instanceof EntityPlayer)
+        {
+            ChowTime.harvestXP = ChowTime.saveData.getInteger("harvestXP" + ((EntityPlayer) event.entity).getCommandSenderName());
+            ChowTime.harvestLVL = ChowTime.saveData.getInteger("harvestLVL" + ((EntityPlayer) event.entity).getCommandSenderName());
+        }
     }
-
+    
     @SubscribeEvent
     public void onWorldLoad(WorldEvent.Load event)
     {
@@ -29,21 +41,19 @@ public class EntityEventHandler
             ChowTime.harvestingLVL = new File(ChowTime.dir + File.separator + "ChowTime", "CT" + event.world.getWorldInfo().getWorldName() + ".cfg");
             try
             {
-                if (!ChowTime.harvestingLVL.exists())
-                    ChowTime.harvestingLVL.createNewFile();
+                if (!ChowTime.harvestingLVL.exists()) ChowTime.harvestingLVL.createNewFile();
             }
             catch (IOException e)
             {
                 e.printStackTrace();
             }
         }
-
+        
         if (FMLCommonHandler.instance().getEffectiveSide().isServer())
         {
             try
             {
-                if (ChowTime.harvestingLVL.exists())
-                    ChowTime.saveData = CompressedStreamTools.readCompressed(new FileInputStream(ChowTime.harvestingLVL));
+                if (ChowTime.harvestingLVL.exists()) ChowTime.saveData = CompressedStreamTools.readCompressed(new FileInputStream(ChowTime.harvestingLVL));
             }
             catch (EOFException e)
             {
@@ -55,7 +65,7 @@ public class EntityEventHandler
             }
         }
     }
-
+    
     @SubscribeEvent
     public void onWorldSave(WorldEvent.Save event)
     {
@@ -63,8 +73,12 @@ public class EntityEventHandler
         {
             try
             {
-                if (ChowTime.harvestingLVL.exists())
-                    CompressedStreamTools.writeCompressed(ChowTime.saveData, new FileOutputStream(ChowTime.harvestingLVL));
+                if (ChowTime.harvestingLVL.exists()) CompressedStreamTools.writeCompressed(ChowTime.saveData, new FileOutputStream(ChowTime.harvestingLVL));
+                int i = event.world.playerEntities.size();
+                for(int j = 0; j < i; j++){
+                    ChowTime.saveData.setInteger("harvestXP" + ((EntityPlayer)event.world.playerEntities.get(j)).getCommandSenderName(), ChowTime.harvestXP);
+                    ChowTime.saveData.setInteger("harvestLVL" + ((EntityPlayer)event.world.playerEntities.get(j)).getCommandSenderName(), ChowTime.harvestLVL);    
+                }
             }
             catch (EOFException e)
             {
@@ -74,6 +88,80 @@ public class EntityEventHandler
             {
                 e.printStackTrace();
             }
+        }
+    }
+    
+    @SubscribeEvent
+    public void onEntityUpdate(LivingEvent.LivingUpdateEvent event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            if (event.entityLiving instanceof EntityPlayer)
+            {
+            }
+        }
+    }
+    
+    @SubscribeEvent
+    public void onItemUseStart(PlayerUseItemEvent.Start event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            
+        }
+    }
+    
+    @SubscribeEvent
+    public void onItemUseTick(PlayerUseItemEvent.Tick event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            
+        }
+    }
+    
+    @SubscribeEvent
+    public void onItemUseStopBeforeFinish(PlayerUseItemEvent.Stop event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            
+        }
+    }
+    
+    @SubscribeEvent
+    public void onItemUseFinish(PlayerUseItemEvent.Finish event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            // if (event.item.getItem() instanceof ItemAppleGold ||
+            // event.item.getItem() instanceof ItemFood || event.item.getItem()
+            // instanceof ItemSoup || event.item.getItem() instanceof
+            // ItemFishFood) event.entityPlayer.addChatMessage(new
+            // ChatComponentText("Munch munch munch"));
+        }
+    }
+    
+    @SubscribeEvent
+    public void breakSpeed(PlayerEvent.BreakSpeed event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            if (event.block instanceof BlockDirt && ChowTime.harvestXP < 50) event.setCanceled(true);
+            if (event.block instanceof CropBarley && event.metadata == 7)
+            {
+                ChowTime.harvestXP++;
+                event.entityPlayer.addChatMessage(new ChatComponentText(Integer.toString(ChowTime.harvestXP)));
+            }
+        }
+    }
+    
+    @SubscribeEvent
+    public void harvestCheck(PlayerEvent.HarvestCheck event)
+    {
+        if (FMLCommonHandler.instance().getEffectiveSide().isServer())
+        {
+            
         }
     }
 }
